Resolve app.core dependency blocking oauth extraction
