+++ app/core/src/main/java/com/fsck/k9/preferences/ServerTypeConverter.kt
@@ -1,5 +1,7 @@
 package com.fsck.k9.preferences
 
+import app.k9mail.core.common.mail.Protocols
+
 object ServerTypeConverter {
     @JvmStatic
     fun toServerSettingsType(exportType: String): String = exportType.lowercase()
