+++ app/core/src/test/java/com/fsck/k9/setup/ServerNameSuggesterTest.java
@@ -1,7 +1,7 @@
 package com.fsck.k9.setup;
 
 
-import com.fsck.k9.preferences.Protocols;
+import app.k9mail.core.common.mail.Protocols;
 import org.junit.Before;
 import org.junit.Test;
