+++ feature/autodiscovery/providersxml/src/test/kotlin/app/k9mail/autodiscovery/providersxml/ProvidersXmlDiscoveryTest.kt
@@ -2,14 +2,14 @@ package app.k9mail.autodiscovery.providersxml
 
 import androidx.test.core.app.ApplicationProvider
 import app.k9mail.core.android.testing.RobolectricTest
+import app.k9mail.core.common.oauth.OAuthConfiguration
+import app.k9mail.core.common.oauth.OAuthConfigurationProvider
 import assertk.assertThat
 import assertk.assertions.isEqualTo
 import assertk.assertions.isNotNull
 import assertk.assertions.isNull
 import com.fsck.k9.mail.AuthType
 import com.fsck.k9.mail.ConnectionSecurity
-import com.fsck.k9.oauth.OAuthConfiguration
-import com.fsck.k9.oauth.OAuthConfigurationProvider
 import org.junit.Test
 
 class ProvidersXmlDiscoveryTest : RobolectricTest() {
